Move 5 task integration tests from aprot to tasks package
Tests that only use the public aprot API now live in tasks/ and exercise
the real tasks.Enable() / tasks.EnableWithMeta() code path instead of
the mirror helpers (enableTestTasks / enableTestTasksWithMeta).

Moved tests:
- TestEnableRegistersHandler (was TestEnableTasksRegistersHandler)
- TestEnableWithMetaRegistersType (was TestEnableTasksWithMeta)
- TestEnableRegistersEnum (was TestEnableTasksRegistersEnum)
- TestSharedSubTaskIsOwnerAlwaysFalse
- TestStartSharedTaskIsOwnerTrue

Also removes the readFirstTaskIsOwner helper from shared_task_test.go
(only used by the two moved WebSocket tests).

Closes #76
