+++ shared_task_test.go
@@ -9,7 +9,6 @@ import (
 	"time"
 
 	"github.com/go-json-experiment/json"
-	"github.com/go-json-experiment/json/jsontext"
 )
 
 func TestSharedTaskBasic(t *testing.T) {
@@ -325,32 +324,6 @@ func TestSharedTaskSubErr(t *testing.T) {
 	}
 
 	task.Close()
-}
-
-// Test that EnableTasks registers the CancelTask handler.
-func TestEnableTasksRegistersHandler(t *testing.T) {
-	registry := NewRegistry()
-	registry.Register(&IntegrationHandlers{})
-	registry.RegisterPushEventFor(&IntegrationHandlers{}, NotificationEvent{})
-	enableTestTasks(registry)
-
-	_, ok := registry.Get("taskCancelHandler.CancelTask")
-	if !ok {
-		t.Fatal("expected CancelTask handler to be registered")
-	}
-
-	// Check push events
-	events := registry.PushEvents()
-	found := map[string]bool{}
-	for _, e := range events {
-		found[e.Name] = true
-	}
-	if !found["TaskStateEvent"] {
-		t.Error("expected TaskStateEvent push event")
-	}
-	if !found["TaskUpdateEvent"] {
-		t.Error("expected TaskUpdateEvent push event")
-	}
 }
 
 // Integration test: SharedTask broadcast over WebSocket.
@@ -1258,162 +1231,3 @@ func TestSharedTaskPushOnConnectIsOwner(t *testing.T) {
 	core.closeTask()
 }
 
-// readFirstTaskIsOwner reads messages from a WebSocket until it finds a TaskStateEvent
-// with at least one task, and returns that task's isOwner value.
-func readFirstTaskIsOwner(ws interface{ ReadMessage() (int, []byte, error) }) *bool {
-	for i := 0; i < 20; i++ {
-		_, data, err := ws.ReadMessage()
-		if err != nil {
-			break
-		}
-		if !strings.Contains(string(data), "TaskStateEvent") {
-			continue
-		}
-		var msg struct {
-			Data struct {
-				Tasks []struct {
-					IsOwner bool `json:"isOwner"`
-				} `json:"tasks"`
-			} `json:"data"`
-		}
-		json.Unmarshal(data, &msg)
-		if len(msg.Data.Tasks) > 0 {
-			v := msg.Data.Tasks[0].IsOwner
-			return &v
-		}
-	}
-	return nil
-}
-
-func TestSharedSubTaskIsOwnerAlwaysFalse(t *testing.T) {
-	registry := NewRegistry()
-	handlers := &IntegrationHandlers{}
-	registry.Register(handlers)
-	registry.RegisterPushEventFor(&IntegrationHandlers{}, NotificationEvent{})
-	enableTestTasks(registry)
-
-	server := NewServer(registry)
-	handlers.server = server
-
-	ts := httptest.NewServer(server)
-	defer ts.Close()
-
-	ws1 := connectWS(t, ts)
-	defer ws1.Close()
-
-	ws2 := connectWS(t, ts)
-	defer ws2.Close()
-
-	// Wait for both connections to be registered.
-	time.Sleep(50 * time.Millisecond)
-
-	// Client 1 calls SharedSubTask via RPC — the handler sleeps 500ms,
-	// giving the flush loop time to broadcast TaskStateEvent.
-	req := IncomingMessage{
-		Type:   TypeRequest,
-		ID:     "1",
-		Method: "IntegrationHandlers.RunSharedSubTask",
-		Params: jsontext.Value(`[{"title":"Deploy"}]`),
-	}
-	if err := ws1.WriteJSON(req); err != nil {
-		t.Fatalf("Write failed: %v", err)
-	}
-
-	ws1.SetReadDeadline(time.Now().Add(2 * time.Second))
-	ws2.SetReadDeadline(time.Now().Add(2 * time.Second))
-
-	isOwner1 := readFirstTaskIsOwner(ws1)
-	isOwner2 := readFirstTaskIsOwner(ws2)
-
-	if isOwner1 == nil || isOwner2 == nil {
-		t.Fatal("expected both clients to receive TaskStateEvent")
-	}
-
-	// Both should be false — SharedSubTask tasks are not top-level.
-	if *isOwner1 {
-		t.Error("expected isOwner=false for caller (SharedSubTask is not top-level)")
-	}
-	if *isOwner2 {
-		t.Error("expected isOwner=false for observer (SharedSubTask is not top-level)")
-	}
-}
-
-func TestStartSharedTaskIsOwnerTrue(t *testing.T) {
-	registry := NewRegistry()
-	handlers := &IntegrationHandlers{}
-	registry.Register(handlers)
-	registry.RegisterPushEventFor(&IntegrationHandlers{}, NotificationEvent{})
-	enableTestTasks(registry)
-
-	server := NewServer(registry)
-	handlers.server = server
-
-	ts := httptest.NewServer(server)
-	defer ts.Close()
-
-	ws1 := connectWS(t, ts)
-	defer ws1.Close()
-
-	ws2 := connectWS(t, ts)
-	defer ws2.Close()
-
-	// Wait for both connections to be registered.
-	time.Sleep(50 * time.Millisecond)
-
-	// Client 1 calls StartSharedTask via RPC — the handler sleeps 500ms,
-	// giving the flush loop time to broadcast TaskStateEvent.
-	req := IncomingMessage{
-		Type:   TypeRequest,
-		ID:     "1",
-		Method: "IntegrationHandlers.RunStartSharedTask",
-		Params: jsontext.Value(`[{"title":"Build"}]`),
-	}
-	if err := ws1.WriteJSON(req); err != nil {
-		t.Fatalf("Write failed: %v", err)
-	}
-
-	ws1.SetReadDeadline(time.Now().Add(2 * time.Second))
-	ws2.SetReadDeadline(time.Now().Add(2 * time.Second))
-
-	isOwner1 := readFirstTaskIsOwner(ws1)
-	isOwner2 := readFirstTaskIsOwner(ws2)
-
-	if isOwner1 == nil || isOwner2 == nil {
-		t.Fatal("expected both clients to receive TaskStateEvent")
-	}
-
-	// Client 1 (caller) should see isOwner=true, client 2 should see false.
-	if !*isOwner1 {
-		t.Error("expected isOwner=true for caller (StartSharedTask is top-level)")
-	}
-	if *isOwner2 {
-		t.Error("expected isOwner=false for observer")
-	}
-}
-
-func TestEnableTasksWithMeta(t *testing.T) {
-	type TaskMeta struct {
-		UserName string `json:"userName,omitempty"`
-		Error    string `json:"error,omitempty"`
-	}
-
-	registry := NewRegistry()
-	registry.Register(&IntegrationHandlers{})
-	registry.RegisterPushEventFor(&IntegrationHandlers{}, NotificationEvent{})
-	enableTestTasksWithMeta[TaskMeta](registry)
-
-	if !registry.TasksEnabled() {
-		t.Fatal("expected tasks to be enabled")
-	}
-
-	metaType := registry.TaskMetaType()
-	if metaType == nil {
-		t.Fatal("expected non-nil meta type")
-	}
-	if metaType.Name() != "TaskMeta" {
-		t.Errorf("expected meta type name 'TaskMeta', got %s", metaType.Name())
-	}
-	if metaType.NumField() != 2 {
-		t.Errorf("expected 2 fields, got %d", metaType.NumField())
-	}
-}
